refactor: remove bookerId and itemId from booking
