+++ src/main/java/ru/practicum/shareit/user/UserMapper.java
@@ -1,5 +1,6 @@
-package ru.practicum.shareit.user.dto;
+package ru.practicum.shareit.user;
 
+import ru.practicum.shareit.user.dto.UserDto;
 import ru.practicum.shareit.user.model.User;
